se agrego generic filter
