+++ src/components/Cards/ProveedorTable/index.js
@@ -4,42 +4,17 @@ import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Table from 'react-bootstrap/Table';
 import Button from 'react-bootstrap/Button';
-import { post, get, onDelete } from 'services';
-import Form from 'react-bootstrap/Form';
-import CurrencyInput from 'react-currency-input';
-import DatePicker from "react-datepicker"
+import { get, onDelete } from 'services';
 import "react-datepicker/dist/react-datepicker.css";
+import GenericFilter from "components/GenericFilter";
 
 const marginBottom = {
     marginBottom: 10
 }
 
-const mtHere = {
-    marginTop: 32
-}
-
-const mr = {
-    ...mtHere,
-    marginRight: 10
-}
-
-const months = ['Enero', 'Febrero', 'Marzo', 'Abril', 'Mayo', 'Junio', 'Julio', 'Agosto', 'Septiembre', 'Octubre', 'Noviembre', 'Diciembre'];
-const days = ['Do', 'Lu', 'Ma', 'Mi', 'Ju', 'Vi', 'Sa'];
-const format = 'dd/MM/yyyy';
-const locale = {
-    localize: {
-        month: n => months[n],
-        day: n => days[n]
-    },
-    formatLong: {}
-};
-
 const ProveedorTable = ({ onHandleChange }) => {
     const [proveedores, setProveedores] = useState([]);
-    const [balance, setBalance] = useState('');
-    const [startDate, setStartDate] = useState();
-    const [endDate, setEndDate] = useState();
-    
+
     useEffect(() => {
         get('Proveedores').then(response => {
             setProveedores(response.data.data)
@@ -61,82 +36,10 @@ const ProveedorTable = ({ onHandleChange }) => {
             });
         })
     }
-
-    const onFilter = () => {
-        let url = 'Proveedores/filter'
-        let data = {}
-
-        if (balance) {
-            data = {
-                balance: parseFloat(balance)
-            }
-        }
-
-        if (startDate && endDate) {
-            url = `Proveedores/filter?fechaInicio=${startDate}&fechaFin=${endDate}`
-        }
-
-        post(url, data).then(response => {
-            setProveedores(response.data.data)
-        });
-    }
-
-    const onClear = () => {
-        get('Proveedores').then(response => {
-            setProveedores(response.data.data)
-        });
-        setBalance('');
-        setStartDate();
-        setEndDate();
-    }
     
     return (
         <>
-        <Card>
-            <Card.Body className={'p-20'}>
-                <Form.Row>
-
-                    <Form.Group as={Col} md="3" controlId={"validationwageAspiratio212"}>
-                        <Form.Label>Balance</Form.Label>
-                        <Form.Control
-                            type="text"
-                            name="balance"
-                            value={balance}
-                            onChange={e => setBalance(e.target.value)}
-                        />
-                    </Form.Group>
-
-                    <Form.Group as={Col} md="3" controlId="validationFormik11">
-                        <Form.Label> Desde </Form.Label>
-                        <DatePicker 
-                            className="form-control"
-                            locale={locale}
-                            dateFormat={format}
-                            selected={startDate} onChange={date => setStartDate(date)}
-                        />
-                    </Form.Group>
-
-                    <Form.Group as={Col} md="3" controlId="validationFormik11">
-                        <Form.Label> Hasta </Form.Label>
-                        <DatePicker 
-                            className="form-control"
-                            locale={locale}
-                            dateFormat={format}
-                            selected={endDate} onChange={date => setEndDate(date)}
-                        />
-                    </Form.Group>
-
-                    <Form.Group as={Col} md="3" controlId="validationFormik11">
-                        <div className="display-flex flex-end">
-                            <Button style={mr} variant="primary" onClick={onClear}>  <i className="fa fa-broom"></i></Button>
-                            <Button style={mtHere} variant="primary" onClick={onFilter}> <i className="fa fa-search"></i></Button>
-                        </div>
-                    </Form.Group>
-
-                </Form.Row>
-                
-            </Card.Body>
-        </Card>
+        <GenericFilter filterByBalance={true} filterByDates={true} endpoint={'Proveedores'} setData={setProveedores}/>
         <br/>
         <Card>
             <Card.Body className={'p-20'}>
