+++ src/router/index.js
@@ -4,8 +4,8 @@ import Login from 'views/login';
 import Register from 'views/register';
 import HomeModule from 'views/home';
 import ProviderModule from 'views/provider'
-import DocumentEntryModule from 'views/DocumentEntry'
-
+import DocumentEntryModule from 'views/documentEntry'
+import Accounting from 'views/accounting'
 
 const Router = () => {
     return (
@@ -15,6 +15,7 @@ const Router = () => {
             <Route path="/documentEntryModule" component={DocumentEntryModule} />
             <Route path="/login" component={Login} />
             <Route path="/register" component={Register} />
+            <Route path="/accounting" component={Accounting} />
         </Switch>
     )
 };
