+++ src/views/accounting/index.js
@@ -1,6 +1,5 @@
 import React, { useState } from 'react';
-import ProveedorTable from 'components/Cards/ProveedorTable';
-import ProveedorForm from 'components/Forms/ProveedorForm';
+import AccountingTable from 'components/Cards/AccountingTable';
 import ToggleDetail from 'components/ToggleDetail';
 import { isBrowser } from "react-device-detect";
 import { connect } from 'react-redux';
@@ -9,7 +8,7 @@ import { SetUser } from "redux/actions/userActions";
 
 const modulePadding =  isBrowser ? '' : '';
 
-const ProviderModule = (props) => {
+const AccountingModule = (props) => {
   const [showDetail, setShowDetail] = useState(true);
   const [action, setAction] = useState('');
   const [id, setId] = useState(0);
@@ -24,16 +23,9 @@ const ProviderModule = (props) => {
     ? 
     (
       <div className={modulePadding}>
-        <ToggleDetail onToggleDetail={setShowDetail} label="Proveedores"/>
+        <ToggleDetail onToggleDetail={setShowDetail} label="Contabilidad"/>
         { 
-          showDetail ? 
-          <div> 
-            <ProveedorTable onHandleChange={onHandleChange}/>
-          </div>
-          : 
-          <div> 
-            <ProveedorForm id={id} action={action} onHandleChange={onHandleChange} />
-          </div>
+          showDetail && <AccountingTable onHandleChange={onHandleChange}/>
         }
       </div>
     )
@@ -52,4 +44,4 @@ const mapDispatchToProps = (dispatch) => {
   }
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(ProviderModule);
+export default connect(mapStateToProps, mapDispatchToProps)(AccountingModule);
